+++ cmd/coin2html/js/body.html
@@ -25,3 +25,6 @@
     <section id="view"></section>
   </section>
 </div>
+
+<script src="src/models.ts"></script>
+<script src="src/views.ts"></script>
