+++ cmd/coin2html/js/head.html
@@ -3,9 +3,6 @@
   <head>
     <meta charset="UTF-8" />
     <title>COIN</title>
-    <script
-      src="https://cdn.jsdelivr.net/npm/d3@7.9.0/dist/d3.min.js"
-      charset="utf-8"
-    ></script>
+    <link href="styles.css" rel="stylesheet" />
   </head>
 </html>
