+++ cmd/coin2html/js/package.json
@@ -1,11 +1,10 @@
 {
   "name": "coin",
-  "version": "1.0.0",
-  "description": "coin companion",
-  "main": "src/index.js",
+  "version": "0.1.0",
+  "description": "coin companion viewer",
   "scripts": {
-    "build": "tsc ",
-    "test": "jest --no-cache --runInBand --env='node' --forceExit",
+    "build": "webpack ",
+    "test": "jest",
     "lint": "prettier --check . && eslint .",
     "autolint": "prettier --write . && eslint --fix ."
   },
@@ -19,11 +18,17 @@
     "d3": "^7.9.0"
   },
   "devDependencies": {
-    "@types/d3": "^7.4.3",
+    "@types/d3": "^7.0.0",
+    "@types/jest": "^27.5.2",
+    "css-loader": "^7.1.2",
     "eslint": "^8.14.0",
-    "jest": "^28.1.0",
+    "html-bundler-webpack-plugin": "^3.17.4",
+    "jest": "^28.1.3",
     "prettier": "^2.6.2",
-    "ts-jest": "^28.0.1",
-    "typescript": "^4.6.4"
+    "ts-jest": "^28.0.8",
+    "ts-loader": "^9.5.1",
+    "typescript": "^4.9.5",
+    "webpack": "^5.94.0",
+    "webpack-cli": "^5.1.4"
   }
 }
