+++ cmd/coin2html/js/spec/commodity.spec.ts
@@ -1,0 +1,4 @@
+import { Commodity } from "../src/models";
+
+test("create commodity", () =>
+  expect(new Commodity("CAD", "Canadian Dollar", 2)).toBeTruthy());
